refactor: 도메인별 Repository 인터페이스 생성 및 인터페이스 다중상속을 통한 SubwayRepository 생성
