+++ src/main/java/wooteco/subway/service/LineService.java
@@ -7,8 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import wooteco.subway.domain.line.Line;
-import wooteco.subway.repository.LineRepository;
-import wooteco.subway.repository.dao.LineDao;
+import wooteco.subway.domain.line.LineRepository;
 import wooteco.subway.service.dto.ServiceDtoAssembler;
 import wooteco.subway.service.dto.line.LineRequestDto;
 import wooteco.subway.service.dto.line.LineResponseDto;
@@ -24,31 +23,31 @@ public class LineService {
 
     @Transactional
     public LineResponseDto create(String name, String color) {
-        Long lineId = lineRepository.save(new Line(name, color));
-        Line line = lineRepository.findById(lineId);
+        Long lineId = lineRepository.saveLine(new Line(name, color));
+        Line line = lineRepository.findLineById(lineId);
         return ServiceDtoAssembler.lineResponseDto(line);
     }
 
     public List<LineResponseDto> findAll() {
-        return lineRepository.findAll()
+        return lineRepository.findLines()
                 .stream()
                 .map(ServiceDtoAssembler::lineResponseDto)
                 .collect(Collectors.toUnmodifiableList());
     }
 
     public LineResponseDto findOne(Long id) {
-        return ServiceDtoAssembler.lineResponseDto(lineRepository.findById(id));
+        return ServiceDtoAssembler.lineResponseDto(lineRepository.findLineById(id));
     }
 
     @Transactional
     public void update(Long id, LineRequestDto lineRequestDto) {
-        Line line = lineRepository.findById(id);
+        Line line = lineRepository.findLineById(id);
         line.update(lineRequestDto.getName(), lineRequestDto.getColor());
-        lineRepository.update(line);
+        lineRepository.updateLine(line);
     }
 
     @Transactional
     public void remove(Long id) {
-        lineRepository.remove(id);
+        lineRepository.removeLine(id);
     }
 }
