+++ src/main/java/wooteco/subway/service/StationService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import wooteco.subway.domain.station.Station;
-import wooteco.subway.repository.StationRepository;
+import wooteco.subway.domain.station.StationRepository;
 import wooteco.subway.service.dto.ServiceDtoAssembler;
 import wooteco.subway.service.dto.station.StationResponseDto;
 
@@ -22,13 +22,13 @@ public class StationService {
 
     @Transactional
     public StationResponseDto create(String name) {
-        Long stationId = stationRepository.save(new Station(name));
-        Station station = stationRepository.findById(stationId);
+        Long stationId = stationRepository.saveStation(new Station(name));
+        Station station = stationRepository.findStationById(stationId);
         return ServiceDtoAssembler.stationResponseDto(station);
     }
 
     public List<StationResponseDto> findAll() {
-        return stationRepository.findAll()
+        return stationRepository.findStations()
                 .stream()
                 .map(ServiceDtoAssembler::stationResponseDto)
                 .collect(Collectors.toUnmodifiableList());
@@ -36,6 +36,6 @@ public class StationService {
 
     @Transactional
     public void remove(Long id) {
-        stationRepository.remove(id);
+        stationRepository.removeStation(id);
     }
 }
